+++ sites/antennasatellite.gr/antennapacific.gr/antennapacific.gr.channels.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<channels>
+  <channel site="antennapacific.gr" site_id="#" lang="el" xmltv_id="ANT1Pacific.gr@SD">ANT1 PACIFIC</channel>
+</channels>
