+++ sites/antennasatellite.gr/antennapacific.gr/readme.md
@@ -1,0 +1,15 @@
+# antennapacific.gr
+
+https://www.antennapacific.gr/el/tvguide.html
+
+### Download the guide
+
+```sh
+npm run grab --- --sites=antennapacific.gr
+```
+
+### Test
+
+```sh
+npm test --- antennapacific.gr
+```
